+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to the [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
+## [1.7.1] - 2020-??-??
+- Deprecate 'decodeBytes' in favor of 'decodeOctets'.
+- Add `Bytes`-oriented encode and decode functions to `Net.Mac`:
+  `boundedBuilderUtf8`, `decodeUtf8Bytes`, and `parserUtf8Bytes`.
+- Add `parserRangeUtf8Bytes` and `parserRangeUtf8BytesLenient` to
+  both `Net.IPv4` and `Net.IPv6` modules.
+
 ## [1.7.0] - 2019-11-05
 - Add `Data` instances for all types.
 - Add `Ix` instances for all address types.